pyproject.toml
[tool.poetry]
name = "fasttext_binary_jawiki"
version = "0.1.0"
description = ""
authors = ["Tetsutaro Maruyama <tetsutaro.maruyama@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
mecab-python3 = "^1.0.4"
neologdn = "^0.5.1"
fasttext = "^0.9.2"
gensim = "^4.1.2"
python-Levenshtein = "^0.12.2"
wikiextractor = "^3.0.6"
tqdm = "^4.64.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "scripts:test"
